Move node config from monad-node to config crate
