+++ monad-node-config/Cargo.toml
@@ -9,9 +9,15 @@ edition = "2021"
 bench = false
 
 [dependencies]
+monad-bls = { workspace = true, optional = true }
+monad-consensus-types = { workspace = true }
 monad-crypto = { workspace = true }
 monad-eth-types = { workspace = true }
+monad-secp = { workspace = true }
 monad-types = { workspace = true }
 
 alloy-primitives = { workspace = true }
 serde = { workspace = true }
+
+[features]
+crypto = ["dep:monad-bls"]
