+++ monad-node-config/src/lib.rs
@@ -1,30 +1,33 @@
 use alloy_primitives::Address;
-use monad_crypto::certificate_signature::PubKey;
-use monad_eth_types::serde::deserialize_eth_address_from_str;
+#[cfg(feature = "crypto")]
+use monad_bls::BlsSignatureCollection;
+#[cfg(feature = "crypto")]
+use monad_consensus_types::checkpoint::Checkpoint;
+use monad_crypto::certificate_signature::{CertificateSignaturePubKey, PubKey};
+use monad_eth_types::{serde::deserialize_eth_address_from_str, EthExecutionProtocol};
+use monad_secp::SecpSignature;
 use serde::Deserialize;
 
+pub use self::{
+    bootstrap::{NodeBootstrapConfig, NodeBootstrapPeerConfig},
+    consensus::NodeConsensusConfig,
+    fullnode::{FullNodeConfig, FullNodeIdentityConfig},
+    network::NodeNetworkConfig,
+    sync_peers::{BlockSyncPeersConfig, StateSyncPeersConfig, SyncPeerIdentityConfig},
+};
+
 mod bootstrap;
-pub use bootstrap::{NodeBootstrapConfig, NodeBootstrapPeerConfig};
-
 mod consensus;
-pub use consensus::NodeConsensusConfig;
-
 mod fullnode;
-pub use fullnode::{FullNodeConfig, FullNodeIdentityConfig};
-
 mod network;
-pub use network::NodeNetworkConfig;
-
 mod sync_peers;
-#[allow(unused_imports)]
-pub use sync_peers::{BlockSyncPeersConfig, StateSyncPeersConfig, SyncPeerIdentityConfig};
 
 #[derive(Debug, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct NodeConfig<P: PubKey> {
-    ////////////////////////////////
+    /////////////////////////////////
     // NODE-SPECIFIC CONFIGURATION //
-    ////////////////////////////////
+    /////////////////////////////////
     pub node_name: String,
     pub network_name: String,
 
@@ -56,3 +59,12 @@ pub struct NodeConfig<P: PubKey> {
     pub chain_id: u64,
     pub consensus: NodeConsensusConfig,
 }
+
+pub type SignatureType = SecpSignature;
+#[cfg(feature = "crypto")]
+pub type SignatureCollectionType =
+    BlsSignatureCollection<CertificateSignaturePubKey<SignatureType>>;
+pub type ExecutionProtocolType = EthExecutionProtocol;
+#[cfg(feature = "crypto")]
+pub type ForkpointConfig = Checkpoint<SignatureCollectionType>;
+pub type MonadNodeConfig = NodeConfig<CertificateSignaturePubKey<SignatureType>>;
