+++ monad-node/Cargo.toml
@@ -4,11 +4,6 @@ description = "Monad consensus node"
 version = "0.1.0"
 edition = "2021"
 
-# this is necessary for criterion bench options
-# https://bheisler.github.io/criterion.rs/book/faq.html#cargo-bench-gives-unrecognized-option-errors-for-valid-command-line-options
-[lib]
-bench = false
-
 [[bin]]
 name = "monad-node"
 bench = false
@@ -23,12 +18,11 @@ monad-crypto = { workspace = true }
 monad-eth-block-policy = { workspace = true }
 monad-eth-block-validator = { workspace = true }
 monad-eth-txpool-executor = { workspace = true }
-monad-eth-types = { workspace = true }
 monad-executor = { workspace = true }
 monad-executor-glue = { workspace = true }
 monad-keystore = { workspace = true }
 monad-ledger = { workspace = true }
-monad-node-config = { workspace = true }
+monad-node-config = { workspace = true, features = ["crypto"] }
 monad-raptorcast = { workspace = true }
 monad-router-filter = { workspace = true, optional = true }
 monad-secp = { workspace = true }
